requirements.txt
streamlit
python-dotenv
newspaper3k
lxml_html_clean
langchain==1.0.5
langchain-classic==1.0.0
langchain-core==1.0.4
langchain-text-splitters==1.0.0
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.41
langchain_openai
langchain_groq
transformers
yt_dlp
git+https://github.com/openai/whisper.git
sentence_transformers
ffmpeg-python
langchain_huggingface
chromadb
langchain-chroma
soundfile
sounddevice
elevenlabs
pypdf